% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{array2ensg}
\alias{array2ensg}
\title{Biomart conversion}
\usage{
array2ensg(
  array_ids,
  array = NULL,
  biomart = mart,
  combine = F,
  df2 = "",
  by.x = array,
  by.y = array,
  all = F
)
}
\arguments{
\item{array_ids}{input}
}
\description{
Biomart conversion
}
