% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{chrStartEndStrand2seq}
\alias{chrStartEndStrand2seq}
\title{Biomart conversion}
\usage{
chrStartEndStrand2seq(
  input,
  cols = c("chromosome_name", "start_position", "end_position", "strand"),
  server = "http://rest.ensembl.org/",
  mask = "-",
  coord_system = "GRCh38"
)
}
\arguments{
\item{input}{either a data.frama or a GRanges object}
}
\description{
Biomart conversion
}
\examples{
chrStartEnd2seq()
}
