% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{ensg2ext_name_biotype}
\alias{ensg2ext_name_biotype}
\title{Biomart conversion}
\usage{
ensg2ext_name_biotype(ensg, biomart = mart, combine = F, df2 = "",
  by.x = "ensembl_gene_id", by.y = "row.names", all = F, title = "",
  ...)
}
\arguments{
\item{ensg}{input}
}
\examples{
ensg2ext_name_biotype()
}
