% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{ensg2hugo}
\alias{ensg2hugo}
\title{Biomart conversion}
\usage{
ensg2hugo(ensg, biomart = mart, combine = F, df2 = "",
  by.x = "ensembl_gene_id", by.y = "ensembl_gene_id", all = F)
}
\arguments{
\item{input}{input}
}
\examples{
ensg2hugo()
}
