% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{ensg2reactome}
\alias{ensg2reactome}
\title{Biomart conversion}
\usage{
ensg2reactome(ensg, biomart = mart, combine = F, df2 = "",
  by.x = "ensembl_gene_id", by.y = "ensembl_gene_id", all = F)
}
\arguments{
\item{ensg}{input}
}
\examples{
ensg2reactome()
}
