% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{ensg2source_status}
\alias{ensg2source_status}
\title{Biomart conversion}
\usage{
ensg2source_status(
  ensg,
  biomart = mart,
  combine = F,
  df2 = "",
  by.x = "ensembl_gene_id",
  by.y = "ensembl_gene_id",
  all = T
)
}
\arguments{
\item{ensg}{input}
}
\description{
Biomart conversion
}
\examples{
ensg2source_status()
}
