% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{enst2entrez}
\alias{enst2entrez}
\title{Biomart conversion}
\usage{
enst2entrez(ENSG_v82, biomart = mart, combine = F, df2 = "",
  by.x = "ensembl_gene_id", by.y = "ensembl_gene_id", all = F)
}
\arguments{
\item{ENSG_v82}{input}
}
\examples{
ensg2entrez()
}
