% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{enst2utr}
\alias{enst2utr}
\title{Biomart conversion}
\usage{
enst2utr(ensg, biomart = mart, combine = F, df2 = "",
  by.x = "ensembl_transcript_id", by.y = "ensembl_transcript_id", all = F)
}
\arguments{
\item{ensg}{input}
}
\examples{
enst2ensg()
}
