% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{exon2ensg.is_constitutive}
\alias{exon2ensg.is_constitutive}
\title{Biomart conversion}
\usage{
exon2ensg.is_constitutive(ense_id, biomart = mart, combine = F, df2 = "",
  by.x = "ensembl_exon_id", by.y = "ensembl_exon_id", all = F)
}
\description{
Biomart conversion
}
\examples{
exon2ensg.is_constitutive()
}
