% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{exon2ensg}
\alias{exon2ensg}
\title{Biomart conversion}
\usage{
exon2ensg(ense_id, biomart = mart, combine = F, df2 = "",
  by.x = "ensembl_exon_id", by.y = "ensembl_exon_id", all = F)
}
\arguments{
\item{input}{input}
}
\examples{
exon2ensg()
}
