% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{ext_name2chr_band}
\alias{ext_name2chr_band}
\title{Biomart conversion}
\usage{
ext_name2chr_band(ext_name, biomart = mart, combine = F, df2 = "",
  by.x = "external_gene_name", by.y = "external_gene_name", all = F)
}
\description{
Biomart conversion
}
\examples{
ext_name2chr_band()
}
