% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{ext_name2chr_TSS_geneStartEnd_strand}
\alias{ext_name2chr_TSS_geneStartEnd_strand}
\title{Biomart conversion}
\usage{
ext_name2chr_TSS_geneStartEnd_strand(ext_name, biomart = mart, combine = F,
  df2 = "", by.x = "external_gene_name", by.y = "external_gene_name",
  all = F)
}
\arguments{
\item{ext_name}{input}
}
\examples{
ext_name2chr_band()
}
