% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{ext_name2enst}
\alias{ext_name2enst}
\title{Biomart conversion}
\usage{
ext_name2enst(ext_name, biomart = mart, combine = F, df2 = "",
  by.x = "external_gene_name", by.y = "external_gene_name", all = F)
}
\arguments{
\item{ext_name}{input}
}
\examples{
ext_name2enst()
}
