% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{ext_name2entrez}
\alias{ext_name2entrez}
\title{Biomart conversion}
\usage{
ext_name2entrez(external_gene_name, biomart = mart, combine = F, df2 = "",
  by.x = "external_gene_name", by.y = "external_gene_name", all = F)
}
\arguments{
\item{input}{input}
}
\examples{
ext_name2entrez()
}
