% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{ext_name2previous_ext_name}
\alias{ext_name2previous_ext_name}
\title{Biomart conversion}
\usage{
ext_name2previous_ext_name(ext_name, hg38.to.hg19 = T, from.mart = NULL,
  to.mart = NULL, combine = F, df2 = "", by.x = "ensembl_gene_id",
  by.y = "ensembl_gene_id", all = F, remove.bracket.info = T)
}
\arguments{
\item{ensg}{input}
}
\examples{
ext_name2previous_ext_name()
}
