% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{ext_name2reactome}
\alias{ext_name2reactome}
\title{Biomart conversion}
\usage{
ext_name2reactome(
  external_gene_name,
  biomart = mart,
  combine = F,
  df2 = "",
  by.x = "external_gene_name",
  by.y = "external_gene_name",
  all = F
)
}
\arguments{
\item{external_gene_name}{input}
}
\description{
Biomart conversion
}
\examples{
ext_name2entrez()
}
