% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{ext_name2wiggleplotr}
\alias{ext_name2wiggleplotr}
\title{Biomart conversion}
\usage{
ext_name2wiggleplotr(
  ext_name,
  biomart = mart,
  combine = F,
  df2 = "",
  by.x = "gene_name",
  by.y = "external_gene_name",
  all = F
)
}
\arguments{
\item{ensg}{input}
}
\description{
Biomart conversion
}
\examples{
ensg2hugo()
}
