% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{getSequence_5UTR}
\alias{getSequence_5UTR}
\title{get 5' UTR sequence}
\usage{
getSequence_5UTR(
  ensg,
  which.mart = mart,
  which.type = "ensembl_gene_id",
  which.seq.type = "5utr",
  fasta.output = ""
)
}
\arguments{
\item{ensg}{input}
}
\description{
get 5' UTR sequence
}
\examples{
getSequence_5UTR()
}
