% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{getSequence_cDNA}
\alias{getSequence_cDNA}
\title{get cDNA sequence}
\usage{
getSequence_cDNA(
  input,
  which.mart = mart,
  which.type = "external_gene_name",
  which.seq.type = "cdna",
  fasta.output = ""
)
}
\arguments{
\item{input}{input}
}
\description{
get cDNA sequence
}
\examples{
getSequence_cDNA()
}
