% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{getSequence_peptide}
\alias{getSequence_peptide}
\title{get peptide sequence}
\usage{
getSequence_peptide(input, which.mart = mart,
  which.type = "external_gene_name", which.seq.type = "peptide",
  fasta.output = "")
}
\arguments{
\item{input}{input}
}
\examples{
getSequence_peptide()
}
