% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{goana.enst}
\alias{goana.enst}
\title{Biomart conversion}
\usage{
\method{goana}{enst}(ensg.diff.exp, ensg.universe = F, biomart = mart,
  pval = "", ont = "", combine = F, df2 = "",
  by.x = "ensembl_gene_id", by.y = "ensembl_gene_id", all = F)
}
\arguments{
\item{ensg.diff.exp}{input}
}
\examples{
goana.ensg()
}
