% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{goana.ext_name}
\alias{goana.ext_name}
\title{Biomart conversion}
\usage{
\method{goana}{ext_name}(ext_name.diff.exp, ensg.universe = F,
  biomart = mart)
}
\arguments{
\item{input}{input}
}
\examples{
goana.ext_name()
}
