% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{goid2ensg_extName_biotype_name_evidence}
\alias{goid2ensg_extName_biotype_name_evidence}
\title{Biomart conversion}
\usage{
goid2ensg_extName_biotype_name_evidence(goid, biomart = mart, combine = F,
  df2 = "", by.x = "go_id", by.y = "go_id", all = F)
}
\arguments{
\item{input}{input}
}
\examples{
goid2ensg_extName_biotype_name_evidence()
}
