% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{hugo2entrez}
\alias{hugo2entrez}
\title{Biomart conversion}
\usage{
hugo2entrez(hgnc, biomart = mart, combine = F, df2 = "",
  by.x = "hgnc_symbol", by.y = "hgnc_symbol", all = F)
}
\arguments{
\item{hgnc}{input}
}
\examples{
hugo2entrez()
}
