% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{hugo2GO}
\alias{hugo2GO}
\title{Biomart conversion}
\usage{
hugo2GO(ext_name, biomart = mart, combine = F, df2 = "",
  by.x = "external_gene_name", by.y = "external_gene_name", all = F)
}
\arguments{
\item{ext_name}{input}
}
\examples{
hugo2GO()
}
