% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{kegga.enst}
\alias{kegga.enst}
\title{Biomart conversion}
\usage{
\method{kegga}{enst}(ensg.diff.exp, ensg.universe = F, biomart = mart)
}
\arguments{
\item{ensg.diff.exp}{input}
}
\examples{
kegga.ensg()
}
