% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{reactome.ensg}
\alias{reactome.ensg}
\title{Biomart conversion}
\usage{
reactome.ensg(ensg.diff.exp, biomart = mart, pvalueCutoff = 0.05,
  readable = T)
}
\description{
Biomart conversion
}
\examples{
reactome.ensg()
}
