% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{reactome2ensg}
\alias{reactome2ensg}
\title{Biomart conversion}
\usage{
reactome2ensg(
  reactome_id,
  biomart = mart,
  combine = F,
  df2 = "",
  by.x = "reactome",
  by.y = "reactome",
  all = F
)
}
\arguments{
\item{reactome_id}{input}
}
\description{
Biomart conversion
}
\examples{
reactome2ensg()
}
