% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{refseq2ensg_ext_name}
\alias{refseq2ensg_ext_name}
\title{Biomart conversion}
\usage{
refseq2ensg_ext_name(
  refseq_ids,
  biomart = mart,
  combine = F,
  df2 = "",
  by.x = "refseq",
  by.y = "refseq",
  all = F
)
}
\arguments{
\item{refseq_ids}{input}
}
\description{
Biomart conversion
}
