% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{uniprot2ensg}
\alias{uniprot2ensg}
\title{Biomart conversion}
\usage{
uniprot2ensg(uniprot_swissprot, useSwissprot = T, biomart = mart,
  combine = F, df2 = "", by.x = "uniprot_swissprot",
  by.y = "uniprot_swissprot", all = F)
}
\arguments{
\item{uniprot_swissprot}{input}
}
\examples{
uniprot2ensg()
}
