% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{entrez2ensg}
\alias{entrez2ensg}
\title{Biomart conversion}
\usage{
entrez2ensg(entrez_ids, biomart = mart, combine = F, df2 = "",
  by.x = "entrezgene", by.y = "entrezgene", all = F)
}
\arguments{
\item{input}{input}
}
\examples{
entrez2ensg()
}
