% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{entrez2ext_name}
\alias{entrez2ext_name}
\alias{entrez2ext_name}
\title{Biomart conversion}
\usage{
entrez2ext_name(entrez_ids, biomart = mart, combine = F, df2 = "",
  by.x = "entrezgene", by.y = "entrezgene", all = F)

entrez2ext_name(entrez_ids, biomart = mart, combine = F, df2 = "",
  by.x = "entrezgene", by.y = "entrezgene", all = F)
}
\arguments{
\item{input}{input}

\item{input}{input}
}
\examples{
entrez2ext_name()
}
