% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{kegga.ext_name}
\alias{kegga.ext_name}
\title{Biomart conversion}
\usage{
\method{kegga}{ext_name}(ext_name.diff.exp, ensg.universe = F,
  biomart = mart)
}
\description{
Biomart conversion
}
\examples{
kegga.ext_name()
}
