% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybiomart.R
\name{uniprot2ensg}
\alias{uniprot2ensg}
\title{Biomart conversion}
\usage{
uniprot2ensg(
  uniprotswissprot,
  useSwissprot = T,
  biomart = mart,
  combine = F,
  df2 = "",
  by.x = "uniprotswissprot",
  by.y = "uniprotswissprot",
  all = F
)
}
\arguments{
\item{uniprotswissprot}{input}
}
\description{
Biomart conversion
}
\examples{
uniprot2ensg()
}
